requirements.txt
streamlit
langgraph
langchain
langchain-openai
python-dotenv
pandas
openpyxl
PyPDF2